:recycle: Refactor: 댓글에서 게시글 도메인의 기능을 사용하는 ACL을 나타내는 클래스가 위치한 패키지 이름 변경
- 댓글이 소비자(클라이언트)로서 다른 도메인에 질의한다고 판단하여 수정함
